+++ package.json
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "@mapbox/polyline": "^0.2.0",
-    "axios": "^0.17.0",
+    "axios": "^0.17.1",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
     "babel-preset-es2015": "^6.24.1",
@@ -28,8 +28,8 @@
     "express": "^4.16.2",
     "express-session": "^1.15.6",
     "jquery": "^3.2.1",
-    "nodemon": "^1.12.1",
     "jsonwebtoken": "^8.1.0",
+    "nodemon": "^1.12.1",
     "passport": "^0.4.0",
     "passport-jwt": "^3.0.1",
     "passport-local": "^1.0.0",
